backend/requirements.txt
daphne==4.0.*
django-cors-headers==3.13.*
django-filter==22.*
django-split-settings==1.2.*
django==4.2.*
sqlparse==0.4.*
djangorestframework-simplejwt==5.2.*
djangorestframework==3.14.*
drf-spectacular==0.25.*
openpyxl==3.0.*
whitenoise==6.*
python-dotenv==1.0.1
django-debug-toolbar==3.8.*
python-dotenv==1.0.1
bandit==1.7.2
model-mommy==2.*
pytest-cov==4.0.*
pytest-django==4.5.*
pytest==7.2.*
pip-audit==2.4.*
wemake-python-styleguide==0.17.*
xenon==0.9.*
dj_database_url==2.3.0
pandas==2.2.3
celery==5.3.6
django-celery-beat==2.5.0
redis==5.0.1
django-celery-results==2.5.1
gunicorn
psycopg2-binary==2.9.*